requirement.txt
numpy
opencv-python==4.5.5.64
pillow
pyyaml
os
glob
scipy